requirements.txt
python-telegram-bot[job-queue]
google-generativeai
mysql-connector-python
APScheduler
python-dotenv